#VR-98 #Added parser for Jump points and switcher to change jump points when floor changes. #Buddy:Vimarsh

+++ src/Modules/DeadZoneMenu/DeadZoneMenuItemRepository.cpp
@@ -43,7 +43,7 @@ namespace Eegeo
                 }
             }
             
-            int DeadZoneMenuItemRepository::GetNumOfDeadZoneMenuItems() const
+            size_t DeadZoneMenuItemRepository::GetNumOfDeadZoneMenuItems() const
             {
                 return m_deadZoneMenuItem.size();
             }
